requirements.txt
streamlit
tensorflow
opencv-python
numpy
pillow
scikit-learn
pandas